[Pal] Drop support for uri==NULL in DkProcessCreate
This feature doesn't seem to make much sense. Also, it was already
broken: Pal/regression/Process3.c assumed uri==NULL means spawning the
process without an executable, but PAL code interpreted this as "spawn
caller's executable again". The documentation doesn't mention this
special case, so it's hard to say which semantics was originally
intended.

+++ Pal/src/host/Linux-SGX/ocall_types.h
@@ -167,7 +167,7 @@ typedef struct {
     unsigned int ms_pid;
     const char* ms_uri;
     int ms_stream_fd;
-    int ms_nargs;
+    size_t ms_nargs;
     const char* ms_args[];
 } ms_ocall_create_process_t;
