+++ Pal/src/host/Linux-SGX/pal_linux.h
@@ -286,7 +286,7 @@ int _DkStreamSecureSave(LIB_SSL_CONTEXT* ssl_ctx, const uint8_t** obuf, size_t* 
 
 #else
 
-int sgx_create_process(const char* uri, int nargs, const char** args, int* stream_fd);
+int sgx_create_process(const char* uri, size_t nargs, const char** args, int* stream_fd);
 
 #ifdef DEBUG
 #ifndef SIGCHLD
